\name{AUC}
\alias{AUC}
\title{AUC}
\description{Function to calculate AUC values for response curves using Trapezoidal integration.}
\usage{AUC(x, y, baseline = 0)}
\arguments{
  \item{x}{the x coordinates of points in the curve}
  \item{y}{the y coordinates of points in the curve}
  \item{baseline}{baseline value of y axis}
}

\value{Returns a data.frame containing all AUC's calculated}

\author{Robert Ernst}





