\name{GSCFRUtils-package}
\alias{GSCFRUtils-package}
\alias{GSCFRUtils}
\docType{package}
\title{R package with utility functions for GSCF.}
\description{R package with utility functions for manipulation, extraction and analysis of data available through GSCF.}
\details{
\tabular{ll}{Package: \tab GSCFRUtils\cr
Type: \tab Package\cr
Title: \tab R package with utility functions for GSCF.\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-05\cr
Author: \tab Robert Ernst\cr
Maintainer: \tab Robert Ernst <robert.ernst@tno.nl>\cr
License: \tab Apache v2.0.1\cr
Depends: \tab stringr\cr}
}
\author{Robert Ernst}

\keyword{ package }


