\name{namedListToDataFrame}
\alias{namedListToDataFrame}
\title{namedListToDataFrame}
\description{Transform a named list to data.frame}
\usage{namedListToDataFrame(list, colName = "token")}
\arguments{
  \item{list}{Give a named list}
  \item{colName}{Number of row containing column names.}
}

\value{Returns the transformed namedList as a data.frame}

\author{Robert Ernst}





