\name{transposeToSubjectLayout}
\alias{transposeToSubjectLayout}
\title{transposeToSubjectLayout}
\description{transpose transcriptomics files to subject layout.}
\usage{transposeToSubjectLayout(filePath)}
\arguments{
  \item{filePath}{filepath of the excel file}
}

\value{Return transposed data.matrix}

\author{Robert Ernst}





