\name{namedListToDataFrame}
\alias{namedListToDataFrame}
\title{namedListToDataFrame}
\description{Transform a named list to table.}
\usage{namedListToDataFrame(list, colName = "token")}
\arguments{
  \item{list}{Give a named list}
  \item{colName}{Number of row containing column names.}
}



\author{Robert Ernst}





