\name{transposeToSubjectLayout}
\alias{transposeToSubjectLayout}
\title{transposeToSubjectLayout}
\description{transpose transcriptomics files to subject layout.}
\usage{transposeToSubjectLayout(filename, dataPath)}
\arguments{
  \item{filename}{excel filename}
  \item{dataPath}{path to excel filename}
}

\value{Return transposed data.matrix}

\author{Robert Ernst}





