\name{transposeToSubjectLayout}
\alias{transposeToSubjectLayout}
\title{transposeToSubjectLayout}
\description{transpose transcriptomics files to subject layout.}
\usage{transposeToSubjectLayout(filename)}
\arguments{
  \item{filename}{full path to the excel file}
}

\value{Return transposed data.matrix}

\author{Robert Ernst}





